src/main/java/ru/netology/Radio.java
package ru.netology;

public class Radio {


    private int currentStation;
    private int currentVolume;

    private int counterStation;

    private int minStation = 0;

    private int maxStation = 10;

    private int minVolume = 0;

    private int maxVolume = 100;



    public Radio(int countStation) {
        this.counterStation = countStation;
        this.maxStation = countStation - 1;
    }

    public int getCurrentStation() {
        return currentStation;
    }

    public int getMaxStation() {
        return maxStation;
    }

    public void setCurrentStation(int newStation) {
        if (newStation > maxStation) {
            return;
        }
        if (newStation < minStation) {
            return;
        }
        currentStation = newStation;
    }

    public void stationNext() {
        if (currentStation >= maxStation) {
            currentStation = minStation;
        } else {
            currentStation = maxStation;
        }
    }


    public void stationPrev() {
        if (currentStation == minStation) {
            currentStation = maxStation;
        } else {
            currentStation = maxStation;
        }
    }
    public int getCurrentVolume() {
        return currentVolume;
    }
    public void setCurrentVolume(int newVolume) {
        currentVolume = newVolume;
    }

    public void increaseVolume() {
        if (currentVolume < maxVolume) {
            currentVolume = currentVolume + 1;
        } else {
            currentVolume = maxVolume;
        }
    }
    public void decreaseVolume() {
        if (currentVolume > minVolume) {
            currentVolume = currentVolume - 1;
        } else {
            currentVolume = minVolume;
        }
    }


}

